requirements.txt
# Core dependencies
PyQt5>=5.15.0
requests>=2.28.0

# Multimodal processing
gTTS>=2.3.0
SpeechRecognition>=3.8.0
pyaudio>=0.2.11
pdfplumber>=0.7.0
pandas>=1.5.0
openpyxl>=3.0.0
Pillow>=9.0.0
numpy>=1.21.0

# Machine Learning & AI
torch>=2.0.0
transformers>=4.30.0
protobuf>=3.20.0
tensorflow>=2.12.0
scipy>=1.9.0
matplotlib>=3.5.0

# Development tools
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950

# Documentation
Sphinx>=4.5.0
sphinx-rtd-theme>=1.0.0